+++ app/src/main/java/com/github/mobile/ui/comment/CommentListAdapter.java
@@ -127,14 +127,7 @@ public class CommentListAdapter extends MultiTypeAdapter {
             textView(0).setTextColor(resources.getColor(R.color.issue_event_normal));
             break;
         case IssueEvent.TYPE_REFERENCED:
-            String commitString;
-            // GitHub sometimes returns a null commit
-            if (event.getCommitId() == null) {
-                commitString = "???????";
-            } else {
-                commitString = event.getCommitId().substring(0,7);
-            }
-            message += String.format(resources.getString(R.string.issue_event_referenced), "<b>" + commitString + "</b>");
+            message += String.format(resources.getString(R.string.issue_event_referenced), "<b>" + event.getCommitId().substring(0,7) + "</b>");
             setText(0, TypefaceUtils.ICON_BOOKMARK);
             textView(0).setTextColor(resources.getColor(R.color.issue_event_normal));
             break;